% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_schedule_columns.r
\name{add_schedule_columns}
\alias{add_schedule_columns}
\title{Add Schedule Columns}
\usage{
add_schedule_columns(df)
}
\description{
This function ...
}
\examples{
add_schedule_columns()
}
\keyword{Autocalibration,EnergyPlus}
