% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_CVRMSE.r
\name{calc_CVRMSE}
\alias{calc_CVRMSE}
\title{Calc Errors: CVRSME}
\usage{
calc_CVRMSE(model, actual)
}
\description{
This function calculates Coefficient of variation of the root mean square error. Metrics for checking model calibration.
From ASHRAE Guideline 14, pg 15, equations 5.4 and 5.5 http://www.eeperformance.org/uploads/8/6/5/0/8650231/ashrae_guideline_14-2002_measurement_of_energy_and_demand_saving.pdf
For hourly analyses the suggested heuristics are < 30% CVRSME.
}
\examples{
calc_CVRMSE()
}
\keyword{Autocalibration,EnergyPlus}
