% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_sections.R
\name{combine_sections}
\alias{combine_sections}
\title{Combine EnergyPlus Input File Sections}
\usage{
combine_sections(run_name)
}
\arguments{
\item{run_name}{Name of the EnergyPlus run}
}
\description{
At ADM we divide the EnergyPlus input files into various sections to make things easier.
}
\examples{
combine_sections()
}
\keyword{Calibration}
\keyword{EnergyPlus,}
\keyword{Parametrics,}
