% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JtoKW.r
\name{JtoKW}
\alias{JtoKW}
\title{Convert to Jules to kW}
\usage{
JtoKW(df)
}
\arguments{
\item{df}{The dataframe of EnergyPlus outputs to be converted}
}
\description{
This function searches for fields in a dataframe representing an EnergyPlus output for units if [J] and then
converts them into units of [kW].
}
\examples{
JtoKW()
}
\keyword{Calibration}
\keyword{Data,}
\keyword{EnergyPlus,Model}
