% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_EP_Output.r
\name{read_EP_Output}
\alias{read_EP_Output}
\title{Read Energy Plus Outputs}
\usage{
read_EP_Output(fileName)
}
\arguments{
\item{fileName}{The name of the file containing billing data to be read in. Must be .csv format.}
}
\description{
This function reads the EnergyPlus output file(s) into R and munges them for use in the autocalibration alogorithms.
}
\examples{
read_EP_Output("parametricModel1_Outputs.csv")

}
\keyword{Calibration}
\keyword{Data,}
\keyword{EnergyPlus,Model}
