% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_parametric_preprocessor.r
\name{run_parametric_preprocessor}
\alias{run_parametric_preprocessor}
\title{Run Parametric PreProcessor}
\usage{
run_parametric_preprocessor(input_file_names, processed_file_name,
  PP_Location = "C:\\\\EnergyPlusV8-5-0\\\\PreProcess\\\\ParametricPreProcessor\\\\parametricpreprocessor")
}
\arguments{
\item{input_file_names}{Input files. Must be .csv format.}

\item{processed_file_name}{Name of output file}

\item{PP_Location}{Location of energy plus parametricpreprocessor.exe}
}
\description{
Takes a list of input file names, including the parametric pre-processor, combines the files, and runs the parametric pre-processor.
}
\examples{
run_parametric_preprocessor
}
\keyword{Calibration}
\keyword{EnergyPlus,}
\keyword{Parametrics,}
