% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_preprocessor_results.r
\name{transfer_preprocessor_results}
\alias{transfer_preprocessor_results}
\title{Transfer Bill PreProcessor Schedule Results}
\usage{
transfer_preprocessor_results(PP_output_file, EP_schedule_file)
}
\arguments{
\item{PP_output_file, EP_schedule_file}{The name of the file containing billing data to be read in. Must be .csv format.}
}
\description{
This function ...
}
\examples{
transfer_preprocessor_results()
}
