% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_RunEPlus_batch.R
\name{update_RunEPlus_batch}
\alias{update_RunEPlus_batch}
\title{Update RunEPlus.bat file}
\usage{
update_RunEPlus_batch()
}
\description{
EnergyPlus is launched using a batch file. Inside are paths which need updating. 
This function uses the working directory to make these updates.
}
\examples{
update_RunEPlus_batch()
}
\keyword{Calibration}
\keyword{EnergyPlus,}
\keyword{Parametrics,}
