% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_schedules.R
\name{update_schedule}
\alias{update_schedule}
\title{Update EnergyPlus Schedules}
\usage{
update_schedule(analytic_schedule_filename)
}
\description{
The analytic models produce schedules. We need to update the EnergyPlus schedule files with the new end-uses.
}
\examples{
update_schedule()
}
\keyword{Calibration}
\keyword{EnergyPlus,}
\keyword{Parametrics,}
